readyup|refactor: Regroup readyup's utils modules by role
Disperses the six modules in `packages/readyup/src/utils/` to directories named for their role: the four domain-free ones to `portable/`, `isPackageInstalled` to `packages/` beside the `resolvePackageRoot` it wraps, and `worseSeverity` to a new `severity/`.

`severity/` also takes `meetsThreshold` from `run/runRdy.ts`, which severs `reporting/reportRdy.ts`'s import from a command module and retires the `meetsThreshold` stub that `kitImportFailure`, `runCommand`, and `stalenessWarning` carried only to satisfy it.

`portable/` now imports nothing outside itself; `jitiImport.ts` previously reached into `utils/` for `toDisplayPath` and `buildInstallCommand`.

Each moved module with a single export takes that export's name: `display-path.ts` becomes `toDisplayPath.ts`, `install-command.ts` becomes `buildInstallCommand.ts`, `resolve-package.ts` becomes `isPackageInstalled.ts`, and `severity.ts` becomes `worseSeverity.ts`.
