+++ packages/readyup/src/layout/layoutEngine.ts
@@ -1,5 +1,5 @@
 import type { Severity, SummaryCounts } from '../kits/types.ts';
-import { pluralizeWithCount } from '../utils/pluralize.ts';
+import { pluralizeWithCount } from '../portable/pluralize.ts';
 import type { Formatter, HeadingLevel, TokenName } from './formatter.ts';
 
 /** Milliseconds below which a line omits its duration, leaving only timings worth reading. */
