+++ packages/readyup/src/portable/pluralize.ts
