+++ packages/readyup/src/run/__tests__/kitImportFailure.unit.test.ts
@@ -16,7 +16,6 @@ vi.mock(import('../../kits/loadRdyKit.ts'), () => ({
 }));
 
 vi.mock(import('../runRdy.ts'), () => ({
-  meetsThreshold: () => true,
   runRdy: mockRunRdy,
 }));
