+++ packages/readyup/src/run/__tests__/runRdy.unit.test.ts
@@ -10,7 +10,7 @@ import type {
   Severity,
   SkipResult,
 } from '../../kits/types.ts';
-import { meetsThreshold, runRdy } from '../runRdy.ts';
+import { runRdy } from '../runRdy.ts';
 
 /*
  * A kit runs as JavaScript, so its functions return whatever their author wrote and its fields hold
@@ -1394,31 +1394,5 @@ describe(runRdy, () => {
 
       expect(report.results[0]?.depth).toBe(0);
     });
-  });
-});
-
-describe(meetsThreshold, () => {
-  it.each([
-    { severity: 'error', threshold: 'error', expected: true },
-    { severity: 'error', threshold: 'warn', expected: true },
-    { severity: 'error', threshold: 'recommend', expected: true },
-    { severity: 'warn', threshold: 'error', expected: false },
-    { severity: 'warn', threshold: 'warn', expected: true },
-    { severity: 'warn', threshold: 'recommend', expected: true },
-    { severity: 'recommend', threshold: 'error', expected: false },
-    { severity: 'recommend', threshold: 'warn', expected: false },
-    { severity: 'recommend', threshold: 'recommend', expected: true },
-  ] as const)('returns $expected for severity=$severity, threshold=$threshold', ({ severity, threshold, expected }) => {
-    expect(meetsThreshold(severity, threshold)).toBe(expected);
-  });
-
-  it('throws on a severity outside the enum', () => {
-    expect(() => meetsThreshold(asSeverity('info'), 'error')).toThrow(
-      'Unknown severity "info". Expected one of: error, warn, recommend.',
-    );
-  });
-
-  it('throws on a threshold outside the enum', () => {
-    expect(() => meetsThreshold('error', asSeverity('info'))).toThrow('Unknown severity threshold "info"');
   });
 });
