+++ packages/readyup/src/run/__tests__/stalenessWarning.unit.test.ts
@@ -28,7 +28,6 @@ vi.mock(import('../../remote/loadRemoteKit.ts'), () => ({
 }));
 
 vi.mock(import('../runRdy.ts'), () => ({
-  meetsThreshold: () => true,
   runRdy: mockRunRdy,
 }));
