+++ packages/readyup/src/run/runRdy.ts
@@ -14,25 +14,15 @@ import type {
   SkippedResult,
 } from '../kits/types.ts';
 import { isFlatChecklist } from '../kits/types.ts';
+import { describeValue } from '../portable/describe-value.ts';
 import { isRecord } from '../portable/isRecord.ts';
-import { describeValue } from '../utils/describe-value.ts';
+import { meetsThreshold } from '../severity/meetsThreshold.ts';
 
 /** Options controlling failure and severity defaults for a run. */
 export interface RunRdyOptions {
   defaultSeverity?: Severity;
   failOn?: Severity;
 }
-
-/**
- * Numeric rank for severity comparison. Lower rank = higher severity.
- *
- * A result "meets or exceeds" a threshold when its rank is <= the threshold's rank.
- */
-const SEVERITY_RANK: Record<Severity, number> = {
-  error: 0,
-  warn: 1,
-  recommend: 2,
-};
 
 /**
  * Severity assigned to a check that is broken rather than failing.
@@ -42,20 +32,6 @@ const SEVERITY_RANK: Record<Severity, number> = {
  * declares.
  */
 const AUTHORING_ERROR_SEVERITY: Severity = 'error';
-
-/**
- * Return true if `severity` is at or above (more severe than or equal to) `threshold`.
- *
- * Throws on a value outside the severity enum. Supplying a validated severity is the caller's
- * responsibility, and the throw is what makes that a contract rather than an assumption: an unranked
- * value compares as `undefined <= n`, which is false, so it would silently exclude the check from
- * both the failure and the reporting thresholds instead of failing loudly.
- */
-export function meetsThreshold(severity: Severity, threshold: Severity): boolean {
-  assertRankedSeverity(severity, 'severity');
-  assertRankedSeverity(threshold, 'severity threshold');
-  return SEVERITY_RANK[severity] <= SEVERITY_RANK[threshold];
-}
 
 /** Resolve the effective severity for a check. */
 function resolveSeverity(check: RdyCheck, defaultSeverity: Severity): Severity {
@@ -350,13 +326,6 @@ export async function runRdy(
   const passed = results.every((r) => !(r.status === 'failed' && meetsThreshold(r.severity, failOn)));
 
   return { results, passed, durationMs };
-}
-
-/** Throw when a severity carries no rank, naming the role it was supplied in. */
-function assertRankedSeverity(severity: Severity, role: string): void {
-  if (!Object.hasOwn(SEVERITY_RANK, severity)) {
-    throw new Error(`Unknown ${role} "${severity}". Expected one of: error, warn, recommend.`);
-  }
 }
 
 /**
